requirements.txt
google-generativeai>=0.8.3
pillow
tqdm
requests
edge-tts
asyncio
